[BW]: [NIP] Generally applicaton_delegate (and lower level callback interfaces) extended to notify bts client about starting mail message transmission (what is needed to fix GUI crashes related to GUI close while transmitting messages).

+++ include/bts/addressbook/contact.hpp
@@ -107,7 +107,7 @@ namespace bts { namespace addressbook {
 
   struct wallet_identity0 : public contact
   { //DLNFIX can we derive wallet_identity from wallet_contact instead to avoid code duplication?
-      wallet_identity0() : mining_effort(0.2) {}
+      wallet_identity0() : mining_effort(static_cast<float>(0.2)) {}
       std::string get_full_name() const { return first_name + " " + last_name; }
       std::string get_display_name() const
       {
